requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-pinecone
langgraph
python-dotenv
httpx
pydantic